+++ include/game/game.h
@@ -19,11 +19,13 @@
 #include "level.h"
 #include "data.h"
 #include "calculator.h"
+#include "modal.h"
 class Camera;
 class Megahack;
 class Menu;
 class Level;
 class Calculator;
+class Modal;
 
 class Game {
 private:
@@ -41,6 +43,7 @@ private:
     Megahack* megahack = nullptr;
     Menu* menu = nullptr;
     Level* level = nullptr;
+    Modal* modal = nullptr;
     GameState state;
 
     Uint8 transitionOpacity = 0;
@@ -67,6 +70,9 @@ public:
     void update();
     void render();
 
+    void openModal(GameState state);
+    void closeModal();
+
     static void delay(int ms, std::function<void()> callback);
 
     GameData* getData() const;
@@ -78,6 +84,7 @@ public:
     GameState getState() const;
     unsigned int getTransitionDuration(unsigned int ms) const;
 
+    void setState(GameState state);
     void setState(GameState state, unsigned int ms, std::function<void()> then);
 };